added and fixed most warnings

+++ app/chat_server/server.cpp
@@ -50,7 +50,7 @@ int main(int argc, char* argv[])
 
     // Run the I/O service on the requested number of threads
     std::vector<std::thread> v;
-    v.reserve(threads - 1);
+    v.reserve(static_cast<long unsigned>(threads - 1));
     for (auto i = threads - 1; i > 0; --i)
         v.emplace_back(
           [&ioc] {
